feat: Move supercacheservice to the common module
