+++ server/common/common-service/src/main/java/io/holoinsight/server/common/config/ProdLog.java
@@ -1,63 +1,62 @@
 /*
  * Copyright 2022 Holoinsight Project Authors. Licensed under Apache-2.0.
  */
-package io.holoinsight.server.home.common.util;
+package io.holoinsight.server.common.config;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import lombok.extern.slf4j.Slf4j;
 
 /**
  *
  * @author jsy1001de
  * @version 1.0: ProdLog.java, v 0.1 2022年02月24日 5:18 下午 jinsong.yjs Exp $
  */
+@Slf4j
 public class ProdLog {
 
-  private static final Logger logger = LoggerFactory.getLogger(ProdLog.class);
 
   public static void info(String who, String what, String how) {
-    logger.info("[" + who + "],[" + what + "],[" + how + "]");
+    log.info("[" + who + "],[" + what + "],[" + how + "]");
   }
 
   public static void info(String what) {
-    logger.info("[" + what + "]");
+    log.info("[" + what + "]");
   }
 
   public static void info(String who, String what) {
-    logger.info("[" + who + "],[" + what + "]");
+    log.info("[" + who + "],[" + what + "]");
   }
 
   public static void debug(String who, String what, String how) {
-    logger.info("[" + who + "],[" + what + "],[" + how + "]");
+    log.info("[" + who + "],[" + what + "],[" + how + "]");
   }
 
   public static void debug(String who, String what) {
-    logger.info("[" + who + "],[" + what + "]");
+    log.info("[" + who + "],[" + what + "]");
   }
 
   public static void warn(String who, String what, Exception e) {
     String sb = "[" + who + "],[" + what + "],[" + e.getMessage() + "]";
-    logger.warn(sb);
+    log.warn(sb);
   }
 
   public static void warn(String who, String what, String how) {
-    logger.warn("[" + who + "],[" + what + "],[" + how + "]");
+    log.warn("[" + who + "],[" + what + "],[" + how + "]");
   }
 
   public static void warn(String who, String what) {
-    logger.warn("[" + who + "],[" + what + "]");
+    log.warn("[" + who + "],[" + what + "]");
   }
 
   public static void error(Object who) {
-    logger.error("[" + who + "]");
+    log.error("[" + who + "]");
   }
 
   public static void error(Object who, String why, Exception e) {
-    logger.error("[" + who + "],[" + why + "]", e);
+    log.error("[" + who + "],[" + why + "]", e);
   }
 
   public static void error(Object who, Throwable e) {
-    logger.error("[" + who + "],[--]", e);
+    log.error("[" + who + "],[--]", e);
   }
 
   /**
@@ -80,6 +79,6 @@ public class ProdLog {
     sb.append(success).append("],[");
     sb.append(count).append("],[");
     sb.append(cost).append("ms]");
-    logger.info(sb.toString());
+    log.info(sb.toString());
   }
 }
