+++ server/common/common-service/src/main/java/io/holoinsight/server/common/config/ScheduleLoadTask.java
@@ -1,7 +1,7 @@
 /*
  * Copyright 2022 Holoinsight Project Authors. Licensed under Apache-2.0.
  */
-package io.holoinsight.server.home.common.util;
+package io.holoinsight.server.common.config;
 
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.concurrent.BasicThreadFactory;
