+++ server/common/common-service/src/main/java/io/holoinsight/server/common/service/CommonServiceAutoConfiguration.java
@@ -25,4 +25,10 @@ public class CommonServiceAutoConfiguration {
   public MetaDataDictValueService metaDataDictValueService() {
     return new MetaDataDictValueServiceImpl();
   }
+
+  @Bean
+  @ConditionalOnMissingBean
+  public SuperCacheService superCacheService() {
+    return new SuperCacheService();
+  }
 }
