+++ server/common/common-service/src/main/java/io/holoinsight/server/common/service/SuperCache.java
@@ -1,7 +1,7 @@
 /*
  * Copyright 2022 Holoinsight Project Authors. Licensed under Apache-2.0.
  */
-package io.holoinsight.server.home.biz.common;
+package io.holoinsight.server.common.service;
 
 
 import io.holoinsight.server.common.dao.entity.MetaDataDictValue;
