+++ server/home/home-service/src/main/java/io/holoinsight/server/home/biz/service/ClusterSchedulerTask.java
@@ -4,9 +4,9 @@
 package io.holoinsight.server.home.biz.service;
 
 import io.holoinsight.server.common.AddressUtil;
+import io.holoinsight.server.common.config.ScheduleLoadTask;
 import io.holoinsight.server.home.common.util.CLUSTER_ROLE_CONST;
-import io.holoinsight.server.home.common.util.ProdLog;
-import io.holoinsight.server.home.common.util.ScheduleLoadTask;
+import io.holoinsight.server.common.config.ProdLog;
 import io.holoinsight.server.home.common.util.StringUtil;
 import io.holoinsight.server.home.dal.model.dto.ClusterDTO;
 import org.springframework.beans.factory.annotation.Autowired;
