+++ server/home/home-task/src/main/java/io/holoinsight/server/home/task/MonitorTaskManager.java
@@ -11,8 +11,8 @@ import io.holoinsight.server.home.biz.service.ClusterTaskService;
 import io.holoinsight.server.home.biz.service.impl.ClusterTaskServiceImpl;
 import io.holoinsight.server.home.common.util.CLUSTER_ROLE_CONST;
 import io.holoinsight.server.home.common.util.Debugger;
-import io.holoinsight.server.home.common.util.ProdLog;
-import io.holoinsight.server.home.common.util.ScheduleLoadTask;
+import io.holoinsight.server.common.config.ProdLog;
+import io.holoinsight.server.common.config.ScheduleLoadTask;
 import io.holoinsight.server.home.dal.model.ClusterTask;
 import io.holoinsight.server.home.dal.model.dto.ClusterDTO;
 import org.springframework.beans.factory.annotation.Autowired;
