+++ server/home/home-web/pom.xml
@@ -30,6 +30,11 @@
             <artifactId>home-alert</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>io.holoinsight.server</groupId>
+            <artifactId>common-service</artifactId>
+            <version>${project.version}</version>
+        </dependency>
 
         <dependency>
             <groupId>org.springframework.boot</groupId>
