+++ server/home/home-web/src/main/java/io/holoinsight/server/home/web/AppInitListener.java
@@ -4,9 +4,8 @@
 package io.holoinsight.server.home.web;
 
 import io.holoinsight.server.home.alert.service.task.coordinator.AlertClusterService;
-import io.holoinsight.server.home.biz.common.SuperCacheService;
+import io.holoinsight.server.common.service.SuperCacheService;
 import io.holoinsight.server.home.biz.service.ClusterSchedulerTask;
-import io.holoinsight.server.home.common.util.ScheduleLoadTask;
 import io.holoinsight.server.home.common.util.cache.local.LocalCacheManage;
 import io.holoinsight.server.home.task.MonitorTaskManager;
 import org.springframework.beans.factory.InitializingBean;
