+++ server/home/home-web/src/main/java/io/holoinsight/server/home/web/controller/QueryFacadeImpl.java
@@ -45,10 +45,8 @@ import io.holoinsight.server.storage.common.model.query.Topology;
 import io.holoinsight.server.storage.common.model.query.TraceBrief;
 import io.holoinsight.server.storage.common.model.query.VirtualComponent;
 import io.holoinsight.server.storage.common.model.specification.sw.Trace;
-import io.holoinsight.server.storage.grpc.trace.ID;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
